+++ src/lib.rs
@@ -4,8 +4,9 @@
 
 use crate::_private::NonExhaustive;
 use crate::menuitem::{MenuItem, Separator};
+use rat_popup::PopupStyle;
 use ratatui::prelude::Style;
-use ratatui::widgets::Block;
+use std::fmt::Debug;
 use std::ops::Range;
 
 pub mod menubar;
@@ -102,7 +103,9 @@ pub struct MenuStyle {
     pub right: Option<Style>,
     pub select: Option<Style>,
     pub focus: Option<Style>,
-    pub popup_block: Option<Block<'static>>,
+
+    pub popup: PopupStyle,
+
     pub non_exhaustive: NonExhaustive,
 }
 
@@ -116,14 +119,14 @@ impl Default for MenuStyle {
             right: None,
             select: None,
             focus: None,
-            popup_block: None,
+            popup: Default::default(),
             non_exhaustive: NonExhaustive,
         }
     }
 }
 
 /// Trait for the structural data of the MenuBar.
-pub trait MenuStructure<'a> {
+pub trait MenuStructure<'a>: Debug {
     /// Main menu.
     fn menus(&'a self, menu: &mut MenuBuilder<'a>);
     /// Submenus.